requirements.txt
cython
numpy
scipy
scikit-learn
requests
uncurl-seq
umap-learn
mnnpy
git+git://github.com/yjzhang/simplex_sample.git
